V 1.3.1

+++ pso4pids/src/main/java/com/DaCodingBeast/pso4pidfs/Arm/ArmAngle.kt
@@ -4,7 +4,7 @@ import CommonUtilities.AngleRange
 
 class ArmAngle (val motor: Hardware.Motor, private val angleOffset: Double) {
     fun findAngle(encoder: Int): Double {
-        val angle = AngleRange.wrap(encoder * (2 * Math.PI / motor.encoderTicksPerRotation))
+        val angle = AngleRange.wrap(encoder.toDouble() * (2 * Math.PI / motor.encoderTicksPerRotation))
         return AngleRange.wrap(angle + angleOffset)
     }
 }
